Resolve db files import confilcts
